Consolidate greyVest theme components into GV library
