+++ src/themes/greyVest/index.js
@@ -1,9 +1,11 @@
 import { defaultProps } from 'recompose'
+import React from 'react'
 
-// components exported from GreyVest component library
+// components exported from component library
 import Box from '../../greyVest/Box'
 import Button from '../../greyVest/Button'
 import Checkbox from '../../greyVest/Checkbox'
+import DateInput from '../../greyVest/DateInput'
 import Fonts from '../../greyVest/Fonts'
 import Icon from '../../greyVest/Icon'
 import ListItem from '../../greyVest/ListItem'
@@ -13,34 +15,38 @@ import RadioList from '../../greyVest/RadioList'
 import Select from '../../greyVest/Select'
 import Style from '../../greyVest/Style'
 import Table from '../../greyVest/Table'
+import Tag from '../../greyVest/Tag'
 import TextHighlight from '../../greyVest/TextHighlight'
 import TextInput from '../../greyVest/TextInput'
 
-//components used only for GreyVest theme
-import DateInput from './DateInput'
-import FilterListItem from './FilterListItem'
+//components used only for theme
+import FilterAdderModalPicker from './FilterAdderModalPicker'
 import MissingTypeComponent from './MissingTypeComponent'
-import ModalPicker from './ModalPicker'
-import Tag from './Tag'
+import PickerItem from './PickerItem'
 
 export default {
-  'FilterAdder.ModalPicker': ModalPicker,
+  'FilterAdder.ModalPicker': FilterAdderModalPicker,
   Box,
   Button,
   Checkbox,
   DateInput,
-  Fonts,
+  Globals: ({ children }) => (
+    <>
+      <Fonts />
+      <Style />
+      {children}
+    </>
+  ),
   Icon,
   TextInput,
   ListItem,
-  PickerItem: FilterListItem,
+  PickerItem,
   NumberInput: defaultProps({ type: 'number' })(TextInput),
   MissingTypeComponent,
   Modal,
   PagerItem,
   RadioList,
   Select,
-  Style,
   Table,
   Tag,
   TextHighlight,
